#47 Moving filtering spam logic to the email validator service
